chore(demo): consolidate NLP demo into a single serve_nlp.py flow
The old snapshot-only server (serve_nlp.py, test_webapp_nlp.py,
explainer_snapshot.pkl, Dockerfile) is dropped in favor of the more
capable live-model flow, previously serve_nlp_ext.py: it supports
configurable models/datasets, dual attribution backends, and custom
model builders that the old script never gained.

- Rename serve_nlp_ext.py -> serve_nlp.py, Dockerfile.nlp_ext -> Dockerfile,
  demo/nlp_ext_cache/ -> demo/nlp_cache/
- Collapse docker-compose.yml to one shapash-nlp service, unify on port 8050
- Update README.md to document only the remaining flow
